requirements.txt
--index-url https://pypi.tuna.tsinghua.edu.cn/simple
--trusted-host pypi.tuna.tsinghua.edu.cn

numpy>=1.24.3
ffmpeg-python>=0.2.0
tqdm>=4.65.0
typing-extensions>=4.7.1
opencv-python>=4.8.0
Pillow>=10.0.0
imagehash>=4.3.1 